hotkey: Disable left arrow hotkey when view has unreads.
On pressing left arrow key while in a message view to edit
the last sent message, if the user has unread messages in the
view older than the last sent message, directly using the hotkey
to edit the last sent message will mark those unreads as read.
Hence, disable the hotkey and warn the user in such a case.
Fixes #17737.